@@ -21,6 +21,7 @@
     "@sentry/tracing": "^7.36.0",
     "@stripe/stripe-js": "^1.42.0",
     "@types/grecaptcha": "^3.0.4",
+    "@types/react-router": "^5.1.20",
     "axios": "^1.5.1",
     "classnames": "^2.3.1",
     "core-js": "^3.25.5",
@@ -31,15 +32,15 @@
     "netlify-identity-widget": "^1.9.2",
     "normalize.css": "^8.0.1",
     "qs": "^6.11.0",
-    "react": "^17.0.0",
+    "react": "^18.2.0",
     "react-autosuggest": "^10.1.0",
-    "react-dom": "^17.0.0",
+    "react-dom": "^18.2.0",
     "react-modal": "^3.16.1",
     "react-number-format": "^5.1.2",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^5.3.4",
     "react-transition-group": "^4.4.1",
     "react-virtualized-auto-sizer": "^1.0.7",
-    "react-window": "^1.8.6",
+    "react-window": "^1.8.7",
     "zustand": "^4.1.5"
   },
   "devDependencies": {
@@ -47,6 +48,7 @@
     "@babel/plugin-proposal-class-properties": "^7.16.7",
     "@babel/preset-env": "^7.20.2",
     "@babel/preset-react": "^7.18.6",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.11",
     "@sentry/webpack-plugin": "^1.20.0",
     "@types/autoprefixer": "^10.2.0",
     "@types/babel__core": "^7.1.19",
@@ -65,12 +67,12 @@
     "@types/netlify-identity-widget": "^1.9.3",
     "@types/prettier": "^2.7.1",
     "@types/qs": "^6.9.7",
-    "@types/react": "^17.0.67",
+    "@types/react": "^18.2.25",
     "@types/react-autosuggest": "^10.1.5",
-    "@types/react-dom": "^16.9.0",
+    "@types/react-dom": "^18",
     "@types/react-modal": "^3.13.1",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/react-transition-group": "^4.4.2",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-transition-group": "^4.4.7",
     "@types/react-virtualized-auto-sizer": "^1.0.1",
     "@types/react-window": "^1.8.5",
     "@types/webpack": "^5.28.0",
@@ -100,8 +102,8 @@
     "mini-css-extract-plugin": "^2.6.1",
     "postcss-loader": "^7.3.3",
     "prettier": "^2.7.1",
-    "react-hot-loader": "^4.13.0",
-    "simple-zustand-devtools": "^1.0.1",
+    "react-refresh": "^0.14.0",
+    "simple-zustand-devtools": "^1.1.0",
     "svg-react-loader": "^0.4.6",
     "ts-loader": "^9.4.1",
     "typescript": "^4.8.4",
