@@ -1,7 +1,3 @@
 import React from 'react';
 
-declare module 'react' {
-  interface ImgHTMLAttributes<T> extends React.HTMLAttributes<T> {
-    loading?: 'auto' | 'eager' | 'lazy';
-  }
-}
+declare module 'react' {}
